crew-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
crewai==0.80.0
crewai-tools==0.14.0
sse-starlette==1.8.2
pydantic==2.6.1
python-dotenv==1.0.1
